requirements.txt
pytest~=8.3.4
pyzmq~=26.2.0
SQLAlchemy~=2.0.36
pip~=24.3.1
Jinja2~=3.1.5
colorama~=0.4.6
typing_extensions~=4.12.2
pluggy~=1.5.0
iniconfig~=2.0.0
packaging~=24.2